+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RoadRunnerStuff/tuning/ManualFeedbackTuner.java
@@ -4,11 +4,7 @@ import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.MecanumDrive;
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.SparkFunOTOSDrive;
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.TankDrive;
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.ThreeDeadWheelLocalizer;
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.TwoDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.Robot.Subsystems.SparkFunOTOSDrive;
 
 public final class ManualFeedbackTuner extends LinearOpMode {
     public static double DISTANCE = 64;
@@ -26,48 +22,6 @@ public final class ManualFeedbackTuner extends LinearOpMode {
                                 .lineToX(DISTANCE)
                                 .lineToX(0)
                                 .build());
-            }
-        } else if (TuningOpModes.DRIVE_CLASS.equals(MecanumDrive.class)) {
-            MecanumDrive drive = new MecanumDrive(hardwareMap, new Pose2d(0, 0, 0));
-            
-            if (drive.localizer instanceof TwoDeadWheelLocalizer) {
-                if (TwoDeadWheelLocalizer.PARAMS.perpXTicks == 0 && TwoDeadWheelLocalizer.PARAMS.parYTicks == 0) {
-                    throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
-                }
-            } else if (drive.localizer instanceof ThreeDeadWheelLocalizer) {
-                if (ThreeDeadWheelLocalizer.PARAMS.perpXTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par0YTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par1YTicks == 1) {
-                    throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
-                }
-            }
-            waitForStart();
-
-            while (opModeIsActive()) {
-                Actions.runBlocking(
-                    drive.actionBuilder(new Pose2d(0, 0, 0))
-                            .lineToX(DISTANCE)
-                            .lineToX(0)
-                            .build());
-            }
-        } else if (TuningOpModes.DRIVE_CLASS.equals(TankDrive.class)) {
-            TankDrive drive = new TankDrive(hardwareMap, new Pose2d(0, 0, 0));
-
-            if (drive.localizer instanceof TwoDeadWheelLocalizer) {
-                if (TwoDeadWheelLocalizer.PARAMS.perpXTicks == 0 && TwoDeadWheelLocalizer.PARAMS.parYTicks == 0) {
-                    throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
-                }
-            } else if (drive.localizer instanceof ThreeDeadWheelLocalizer) {
-                if (ThreeDeadWheelLocalizer.PARAMS.perpXTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par0YTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par1YTicks == 1) {
-                    throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
-                }
-            }
-            waitForStart();
-
-            while (opModeIsActive()) {
-                Actions.runBlocking(
-                    drive.actionBuilder(new Pose2d(0, 0, 0))
-                            .lineToX(DISTANCE)
-                            .lineToX(0)
-                            .build());
             }
         } else {
             throw new RuntimeException();
