+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RoadRunnerStuff/tuning/SplineTest.java
@@ -5,8 +5,7 @@ import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.MecanumDrive;
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.SparkFunOTOSDrive;
+import org.firstinspires.ftc.teamcode.Robot.Subsystems.SparkFunOTOSDrive;
 
 public final class SplineTest extends LinearOpMode {
     @Override
