+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RoadRunnerStuff/tuning/otos/OTOSAngularScalar.java
@@ -5,7 +5,7 @@ import com.acmerobotics.roadrunner.Rotation2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.SparkFunOTOSDrive;
+import org.firstinspires.ftc.teamcode.Robot.Subsystems.SparkFunOTOSDrive;
 @TeleOp
 public class OTOSAngularScalar extends LinearOpMode {
