+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/AxonTeleopTest.java
@@ -3,14 +3,12 @@ package org.firstinspires.ftc.teamcode.Testing;
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.AnalogInput;
-import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.teamcode.Commands.Drive;
-import org.firstinspires.ftc.teamcode.RoadRunnerStuff.SparkFunOTOSDrive;
+import org.firstinspires.ftc.teamcode.Robot.Subsystems.SparkFunOTOSDrive;
 import org.firstinspires.ftc.teamcode.Robot.Subsystems.LMECSubsystem;
 
 @TeleOp(name="axon teleop test", group="LeoGoodCode")
@@ -18,7 +16,8 @@ import org.firstinspires.ftc.teamcode.Robot.Subsystems.LMECSubsystem;
 public class AxonTeleopTest extends LinearOpMode {
 
     AnalogInput analogInput;
-    Drive drive = new Drive(hardwareMap, new SparkFunOTOSDrive(hardwareMap, new Pose2d(new Vector2d(3, 3),3)), new LMECSubsystem(hardwareMap, telemetry));
+
+    Drive drive = new Drive(new SparkFunOTOSDrive(hardwareMap, new Pose2d(new Vector2d(3, 3),3)), new LMECSubsystem(hardwareMap));
 
     @Override
     public void runOpMode() {
